+++ metaflow-netflixext/metaflow_extensions/netflixext/plugins/conda/conda_environment.py
@@ -202,7 +202,7 @@ class CondaEnvironment(MetaflowEnvironment):
                 "echo 'Bootstrapping environment ...'",
                 'python -m %s.remote_bootstrap "%s" "%s" %s %s %s'
                 % (
-                    "metaflow_extensions.nflx.plugins.conda",
+                    "metaflow_extensions.netflixext.plugins.conda",
                     self._flow.name,
                     step_name,
                     arg1,
