+++ metaflow-netflixext/metaflow_extensions/netflixext/plugins/conda/resolvers/conda_lock_resolver.py
@@ -37,7 +37,7 @@ from . import Resolver
 from .pip_resolver import _FAKE_WHEEL
 
 if TYPE_CHECKING:
-    from metaflow_extensions.nflx.plugins.conda.env_descr import (
+    from metaflow_extensions.netflixext.plugins.conda.env_descr import (
         PackageSpecification,
     )  # noqa
